SearchBox component splitted into searchBoxPage & SearchBox, where searchBoxPage concerned about the conecting & recieving props from the store & SearchBox component concerned about rendering the component.This will make the unit testing really really easy
